added model view controller architecture
