+++ routes/tourRoutes.js
@@ -1,0 +1,17 @@
+const express = require('express');
+const { getAllTours, createTour, getTour, updateTour, deleteTour } = require('../controllers/tourController');
+
+const router = express.Router(); //middleware
+
+router
+	.route('/')
+	.get(getAllTours)
+	.post(createTour);
+
+router
+	.route('/:id')
+	.get(getTour)
+	.patch(updateTour)
+	.delete(deleteTour);
+
+module.exports = router;
