+++ routes/userRoutes.js
@@ -1,0 +1,17 @@
+const express = require('express');
+const userController = require('../controllers/userController');
+
+const router = express.Router(); //middleware
+
+
+router
+	.route('/')
+	.get(userController.getAllUsers)
+	.post(userController.createUser);
+router
+	.route('/:id')
+	.get(userController.getUser)
+	.patch(userController.updateUser)
+	.delete(userController.deleteUser);
+
+module.exports = router;
